+++ platform_plugin_turnitin/settings/production.py
@@ -29,6 +29,10 @@ def plugin_settings(settings):
         "TURNITIN_TCA_API_KEY", settings.TURNITIN_TCA_API_KEY
     )
 
+    settings.TURNITIN_TCA_REQUIRE_EULA = getattr(settings, "ENV_TOKENS", {}).get(
+        "TURNITIN_TCA_REQUIRE_EULA", settings.TURNITIN_TCA_REQUIRE_EULA
+    )
+
     settings.TURNITIN_SIMILARITY_REPORT_PAYLOAD = getattr(settings, "ENV_TOKENS", {}).get(
         "TURNITIN_SIMILARITY_REPORT_PAYLOAD",
         settings.TURNITIN_SIMILARITY_REPORT_PAYLOAD,